requirements.txt
# requirements.txt
slack_sdk
Flask
python-dotenv
google-generativeai
APScheduler